requirements.txt
asgiref==3.6.0
astroid==2.12.12
autopep8==2.0.0
dill==0.3.6
Django==4.1.7
django-filter==23.1
djangorestframework==3.14.0
isort==5.10.1
lazy-object-proxy==1.8.0
Markdown==3.4.3
mccabe==0.7.0
platformdirs==2.5.4
psycopg2-binary==2.9.6
pycodestyle==2.9.1
pylint==2.15.5
pytz==2023.3
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.6
wrapt==1.14.1
